+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/TagAcceptanceTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.woowacourse.gongseek.article.domain.Category;
 import com.woowacourse.gongseek.auth.application.dto.AccessTokenResponse;
-import com.woowacourse.gongseek.tag.presentation.dto.TagsResponse;
+import com.woowacourse.gongseek.tag.application.dto.TagsResponse;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
