+++ backend/src/test/java/com/woowacourse/gongseek/tag/application/TagServiceTest.java
@@ -10,10 +10,10 @@ import com.woowacourse.gongseek.article.domain.repository.ArticleTagRepository;
 import com.woowacourse.gongseek.member.domain.Member;
 import com.woowacourse.gongseek.member.domain.repository.MemberRepository;
 import com.woowacourse.gongseek.support.IntegrationTest;
+import com.woowacourse.gongseek.tag.application.dto.TagsResponse;
 import com.woowacourse.gongseek.tag.domain.Tag;
 import com.woowacourse.gongseek.tag.domain.Tags;
 import com.woowacourse.gongseek.tag.domain.repository.TagRepository;
-import com.woowacourse.gongseek.tag.presentation.dto.TagsResponse;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
